trend-news-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
feedparser==6.0.11
newspaper3k==0.2.8
openai==1.52.2
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.6.0
numpy==2.1.2
